Take this with a pinch of salt. It is still very much in the alpha stage
=================================
TODO
Comment properly
Sort out the structure for goodness sake!
Double check the residuals and the errors are computed correctly by
each method using aa test file with GNUPlot.

============
Note IOUtil is not really needed here but I put it in out of habit. It
may be useful if the input file has a different format.