Cleaned up. Still a bit to do.
Made it accept files from the commandline instead of hardcoding. This
is a much better way although it might be tedious if I am working with
lots of files so need to think of something for that later. I have
changed the format of input files so it's

length = int
x error = double
y error = double
data_array