Its a bit more versatile.
Still a fair bit too much hard coding going on but it is tricky to get
the balance when there seem to be lots of ways this program could be
used. I am getting more of an idea how to organise it better now
though. For now:

TODO split the plot filewriting into a different class to all the peak
type stuff.